# api/openapi.yaml
openapi: 3.0.3
info:
  title: API de Alcançabilidade Urbana — Cidade de 15 Minutos
  version: 1.0.0
  description: >-
    API REST para consultar índices de alcançabilidade urbana baseados no conceito de Cidade de 15 Minutos.
    Dados de mobilidade urbana fornecidos sob licença ODbL pelos colaboradores do OpenStreetMap (OSM).
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /geocodificar:
    get:
      summary: Busca cidades canônicas no Nominatim
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Termo de busca (3 a 120 caracteres)
          example: "Praia Grande"
      responses:
        '200':
          description: Resultados encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    osm_tipo:
                      type: string
                      example: "relation"
                    osm_id:
                      type: integer
                      example: 298316
                    nome_exibicao:
                      type: string
                      example: "Praia Grande, São Paulo, Brasil"
                    tipo:
                      type: string
                      example: "administrative"
                    ja_processada:
                      type: boolean
                      example: true
                    cidade_id:
                      type: integer
                      nullable: true
                      example: 22
        '400':
          description: Termo de busca inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '502':
          description: Falha de comunicação com o Nominatim
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /vitrine:
    get:
      summary: Consulta a existência de serviços no Overpass (Vitrine)
      parameters:
        - name: osm_tipo
          in: query
          required: true
          schema:
            type: string
            enum: [relation, way]
          example: "relation"
        - name: osm_id
          in: query
          required: true
          schema:
            type: integer
          example: 298285
      responses:
        '200':
          description: Lista de serviços com as quantidades encontradas no município
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    chave:
                      type: string
                      example: "saude"
                    rotulo:
                      type: string
                      example: "Saúde (hospitais)"
                    grupo:
                      type: string
                      example: "Saúde"
                    cor:
                      type: string
                      example: "#e74c3c"
                    padrao:
                      type: boolean
                      example: true
                    quantidade:
                      type: integer
                      example: 12
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '503':
          description: Serviço do OpenStreetMap indisponível ou sobrecarregado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
  /alcancabilidade:
    get:
      summary: Detalha alcançabilidade de um ponto clicado no mapa
      parameters:
        - name: cidade_id
          in: query
          required: true
          schema:
            type: integer
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
        - name: velocidade
          in: query
          required: false
          schema:
            type: number
            minimum: 2.0
            maximum: 6.0
      responses:
        '200':
          description: Detalhes do nó mais próximo e tempos de acesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalhamentoPonto'
        '404':
          description: Ponto fora da área processada da cidade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /comparar:
    get:
      summary: Compara índices de até 5 cidades
      parameters:
        - name: cidades
          in: query
          required: true
          schema:
            type: string
            example: "1,2"
      responses:
        '200':
          description: Lista de dados comparativos das cidades
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    cidade:
                      $ref: '#/components/schemas/CidadeResumida'
                    indices:
                      type: array
                      items:
                        $ref: '#/components/schemas/IndiceCategoria'
                    moreno:
                      $ref: '#/components/schemas/Moreno'
                      nullable: true
        '400':
          description: Mais de 5 cidades solicitadas ou IDs inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /rota:
    get:
      summary: Rota a pé pela malha viária entre dois nós do grafo
      description: >-
        Calcula, via Dijkstra sobre as arestas reais do OpenStreetMap, o
        caminho a pé entre dois nós da cidade e retorna o traçado como
        GeoJSON LineString. Os identificadores 'de' e 'para' são osm_id de
        nós — obtidos, por exemplo, em /alcancabilidade (campos no.osm_id e
        categorias[].servico_mais_proximo.osm_no_id).
      parameters:
        - name: cidade_id
          in: query
          required: true
          schema:
            type: integer
          example: 10
        - name: de
          in: query
          required: true
          schema:
            type: string
          example: "2062713466"
        - name: para
          in: query
          required: true
          schema:
            type: string
          example: "633283351"
        - name: velocidade
          in: query
          required: false
          schema:
            type: number
            minimum: 2.0
            maximum: 6.0
      responses:
        '200':
          description: Rota encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  tempo_min:
                    type: number
                    example: 7.32
                  qtd_nos_caminho:
                    type: integer
                    example: 25
                  geojson:
                    type: object
                    properties:
                      type:
                        type: string
                        example: LineString
                      coordinates:
                        type: array
                        items:
                          type: array
                          items:
                            type: number
                        example: [[-46.4028, -24.0058], [-46.4031, -24.0052]]
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Cidade sem malha gravada ou caminho inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /processamentos:
    post:
      summary: Dispara o processamento de uma nova cidade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                consulta_osm:
                  type: string
                  example: "Santos, São Paulo, Brazil"
                osm_tipo:
                  type: string
                  enum: [relation, way]
                  example: "relation"
                osm_id:
                  type: integer
                  example: 298285
                nome_exibicao:
                  type: string
                  example: "Santos"
                categorias:
                  type: array
                  items:
                    type: string
                  example: ["saude", "educacao"]
      responses:
        '202':
          description: Processamento iniciado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  consulta_osm:
                    type: string
                  status:
                    type: string
                    example: "rodando"
        '200':
          description: Cidade já processada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ja_processada:
                    type: boolean
                    example: true
                  cidade_id:
                    type: integer
                    example: 10
        '400':
          description: Parâmetro inválido ou injeção de parâmetros detectada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '409':
          description: Já existe um processamento em andamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
components:
  schemas:
    CidadeResumida:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: "Praia Grande"
        pais:
          type: string
          example: "Brazil"
        consulta_osm:
          type: string
          example: "Praia Grande, São Paulo, Brazil"
        data_calculo:
          type: string
          format: date-time
          example: "2026-07-03T10:00:00Z"
        qtd_nos:
          type: integer
          example: 6500
        limiar_minutos:
          type: integer
          example: 15
        indice_geral:
          type: number
          example: 78.45
...
    IndiceCategoria:
      type: object
      properties:
        chave:
          type: string
          example: "saude"
        rotulo:
          type: string
          example: "Saúde (hospitais)"
        cor_hex:
          type: string
          example: "#e74c3c"
        tempo_medio_min:
          type: number
          nullable: true
          example: 12.34
        pct_dentro_limiar:
          type: number
          example: 85.20
        indice:
          type: number
          example: 85.20
    DetalhamentoPonto:
      type: object
      properties:
        no:
          type: object
          properties:
            osm_id:
              type: integer
              format: int64
              example: 457812
            lat:
              type: number
            lon:
              type: number
            distancia_m:
              type: number
              example: 15.4
        indice_ponto:
          type: number
          example: 75.0
        categorias:
          type: array
          items:
            type: object
            properties:
              chave:
                type: string
              rotulo:
                type: string
              cor_hex:
                type: string
              tempo_min:
                type: number
                nullable: true
                example: 10.2
              dentro_limiar:
                type: boolean
                example: true
              servico_mais_proximo:
                type: object
                nullable: true
                properties:
                  id:
                    type: integer
                  nome:
                    type: string
                    nullable: true
                  lat:
                    type: number
                  lon:
                    type: number
    Erro:
      type: object
      properties:
        erro:
          type: string
          example: "Cidade não encontrada"
        codigo:
          type: integer
          example: 404
    Moreno:
      type: object
      properties:
        limiar_minutos:
          type: integer
          example: 15
        pct_cobertura_plena:
          type: number
          example: 78.45
        minutos_cidade:
          type: integer
          nullable: true
          example: 22
        tempo_pior_medio:
          type: number
          nullable: true
          example: 18.23
        tempo_pior_mediana:
          type: number
          nullable: true
          example: 16.50
        pct_nos_sem_cobertura:
          type: number
          example: 12.30
        atende_conceito:
          type: boolean
          example: false
        classificacao:
          type: string
          example: "Parcialmente aderente"
        categoria_gargalo:
          type: object
          nullable: true
          properties:
            chave:
              type: string
              example: "saude"
            rotulo:
              type: string
              example: "Saúde (hospitais)"
            cor_hex:
              type: string
              example: "#e74c3c"
        pct_gargalo:
          type: number
          nullable: true
          example: 45.20
        categorias_ausentes:
          type: array
          items:
            type: object
            properties:
              chave:
                type: string
              rotulo:
                type: string
        distribuicao:
          type: array
          items:
            type: object
            properties:
              faixa:
                type: string
                example: "0-5"
              qtd:
                type: integer
                example: 125
...